refactor(sdk): q7e3-F-02 move public types to lib/types.ts
Cut the 6 published types out of the 14-commit accelerator-prover.ts hotspot
into a neutral lib/types.ts; index.ts re-exports them unchanged (barrel
byte-identical — public-contract.test.ts green); accelerator-transport.ts +
its test import from types.ts instead of back-importing from the prover,
killing the 2-way edge. Type-only move; tsc + 41/41 SDK unit tests green.
