+++ packages/sdk/src/lib/accelerator-transport.ts
@@ -1,6 +1,7 @@
 import ky from "ky";
 import ms from "ms";
-import type { AcceleratorProtocol, AcceleratorStatus } from "./accelerator-prover.js";
+// q7e3-F-02: import shared types from the neutral module, not back from the prover (kills the 2-way edge).
+import type { AcceleratorProtocol, AcceleratorStatus } from "./types.js";
 
 /** How long a probed {@link AcceleratorStatus} stays fresh before a re-probe. */
 const STATUS_CACHE_TTL_MS = 10_000;
